Refactor more of bailman
