--- src/main/java/bounty/manager/Constants.java
@@ -1,7 +1,0 @@
-package bounty.manager;
-
-public class Constants {
-
-    // bounty danger level
-    public final static String[] DANGER_LEVEL = { "none", "low", "medium", "high", "very high" };
-}
