+++ src/main/java/bountylib/AssignmentAi.java
@@ -1,4 +1,4 @@
-package bounty.lib;
+package bountylib;
 
 import java.util.Arrays;
 import java.util.List;
